+++ base/clacgv/test/test.ndarray.js
@@ -21,29 +21,29 @@
 // MODULES //
 
 var tape = require( 'tape' );
-var isSameComplex128Array = require( '@stdlib/assert/is-same-complex128array' );
-var Complex128Array = require( '@stdlib/array/complex128' );
-var zlacgv = require( './../lib/ndarray.js' );
+var isSameComplex64Array = require( '@stdlib/assert/is-same-complex64array' );
+var Complex64Array = require( '@stdlib/array/complex64' );
+var clacgv = require( './../lib/ndarray.js' );
 
 
 // TESTS //
 
 tape( 'main export is a function', function test( t ) {
 	t.ok( true, __filename );
-	t.strictEqual( typeof zlacgv, 'function', 'main export is a function' );
+	t.strictEqual( typeof clacgv, 'function', 'main export is a function' );
 	t.end();
 });
 
 tape( 'the function has an arity of 4', function test( t ) {
-	t.strictEqual( zlacgv.length, 4, 'arity of 4' );
+	t.strictEqual( clacgv.length, 4, 'arity of 4' );
 	t.end();
 });
 
 tape( 'the function conjugates elements in a strided array', function test( t ) {
 	var expected;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array([
+	cx = new Complex64Array([
 		0.3, // 1
 		0.1, // 1
 		0.5, // 2
@@ -58,9 +58,9 @@ tape( 'the function conjugates elements in a strided array', function test( t ) 
 		3.0
 	]);
 
-	zlacgv( 4, zx, 1, 0 );
+	clacgv( 4, cx, 1, 0 );
 
-	expected = new Complex128Array([
+	expected = new Complex64Array([
 		0.3,  // 1
 		-0.1, // 1
 		0.5,  // 2
@@ -75,15 +75,15 @@ tape( 'the function conjugates elements in a strided array', function test( t ) 
 		3.0
 	]);
 
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 	t.end();
 });
 
-tape( 'the function supports a `zx` stride', function test( t ) {
+tape( 'the function supports a `cx` stride', function test( t ) {
 	var expected;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array([
+	cx = new Complex64Array([
 		0.1,  // 1
 		0.1,  // 1
 		3.0,
@@ -98,9 +98,9 @@ tape( 'the function supports a `zx` stride', function test( t ) {
 		2.0
 	]);
 
-	zlacgv( 3, zx, 2, 0 );
+	clacgv( 3, cx, 2, 0 );
 
-	expected = new Complex128Array([
+	expected = new Complex64Array([
 		0.1,   // 1
 		-0.1,  // 1
 		3.0,
@@ -115,15 +115,15 @@ tape( 'the function supports a `zx` stride', function test( t ) {
 		2.0
 	]);
 
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 	t.end();
 });
 
-tape( 'the function supports a negative `zx` stride', function test( t ) {
+tape( 'the function supports a negative `cx` stride', function test( t ) {
 	var expected;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array([
+	cx = new Complex64Array([
 		0.1,  // 3
 		0.1,  // 3
 		3.0,
@@ -138,9 +138,9 @@ tape( 'the function supports a negative `zx` stride', function test( t ) {
 		2.0
 	]);
 
-	zlacgv( 3, zx, -2, 4 );
+	clacgv( 3, cx, -2, 4 );
 
-	expected = new Complex128Array([
+	expected = new Complex64Array([
 		0.1,   // 3
 		-0.1,  // 3
 		3.0,
@@ -155,15 +155,15 @@ tape( 'the function supports a negative `zx` stride', function test( t ) {
 		2.0
 	]);
 
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 	t.end();
 });
 
-tape( 'the function supports a `zx` offset', function test( t ) {
+tape( 'the function supports a `cx` offset', function test( t ) {
 	var expected;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array([
+	cx = new Complex64Array([
 		0.1,
 		0.1,
 		3.0,
@@ -178,9 +178,9 @@ tape( 'the function supports a `zx` offset', function test( t ) {
 		2.0   // 3
 	]);
 
-	zlacgv( 3, zx, 1, 3 );
+	clacgv( 3, cx, 1, 3 );
 
-	expected = new Complex128Array([
+	expected = new Complex64Array([
 		0.1,
 		0.1,
 		3.0,
@@ -195,29 +195,29 @@ tape( 'the function supports a `zx` offset', function test( t ) {
 		-2.0  // 3
 	]);
 
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 	t.end();
 });
 
 tape( 'the function returns a reference to the input array', function test( t ) {
 	var out;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array( [ 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0 ] );
+	cx = new Complex64Array( [ 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0 ] );
 
-	out = zlacgv( 4, zx, 1, 0 );
+	out = clacgv( 4, cx, 1, 0 );
 
-	t.strictEqual( out, zx, 'same reference' );
+	t.strictEqual( out, cx, 'same reference' );
 	t.end();
 });
 
 tape( 'if provided an `N` parameter less than or equal to `0`, the function returns the input array unchanged', function test( t ) {
 	var expected;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array( [ 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0 ] );
+	cx = new Complex64Array( [ 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0 ] );
 
-	expected = new Complex128Array([
+	expected = new Complex64Array([
 		1.0,
 		2.0,
 		3.0,
@@ -228,20 +228,20 @@ tape( 'if provided an `N` parameter less than or equal to `0`, the function retu
 		8.0
 	]);
 
-	zlacgv( -1, zx, 1, 0 );
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	clacgv( -1, cx, 1, 0 );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 
-	zlacgv( 0, zx, 1, 0 );
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	clacgv( 0, cx, 1, 0 );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 
 	t.end();
 });
 
 tape( 'the function supports complex access patterns', function test( t ) {
 	var expected;
-	var zx;
+	var cx;
 
-	zx = new Complex128Array([
+	cx = new Complex64Array([
 		0.1,
 		0.1,
 		3.0,
@@ -258,9 +258,9 @@ tape( 'the function supports complex access patterns', function test( t ) {
 		3.0  // 2
 	]);
 
-	zlacgv( 2, zx, -3, 6 );
+	clacgv( 2, cx, -3, 6 );
 
-	expected = new Complex128Array([
+	expected = new Complex64Array([
 		0.1,
 		0.1,
 		3.0,
@@ -277,7 +277,7 @@ tape( 'the function supports complex access patterns', function test( t ) {
 		-3.0  // 2
 	]);
 
-	t.ok( isSameComplex128Array( zx, expected ), 'returns expected value' );
+	t.ok( isSameComplex64Array( cx, expected ), 'returns expected value' );
 
 	t.end();
 });